pyproject.toml
[project]
name = "Langgraph-chatbot-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.23",
    "langchain-core>=0.3.54",
    "langchain-groq>=0.3.2",
    "langgraph~=0.2.26",
    "python-dotenv>=1.1.0",
]
